fix(ci): separate iOS FCM integration gate
